refactor. Globe 파일 수정
