+++ src/components/globe/GlobeCluster.tsx
@@ -14,15 +14,23 @@ interface Props {
 
 export const globeCluster = ({ mapLocation, postsData, mount, flyToLocation, postModalOpen }: Props) => {
   const clusterData = postsData?.slice(5);
-
   pinImages.forEach(({ name, url }) => loadPinImage(mapLocation, name, url));
-
   if (clusterData) {
     const getValue = mapLocation.getSource('pinPoint');
     if (getValue) {
       removeMapLayersAndSource(mapLocation);
     }
 
+    const handleUnclusteredPointClick = async (e: any) => {
+      if (postsData) {
+        const postId = await e.features[0].properties.cluster;
+        const unclusteredData = postsData.filter(item => item.id === postId);
+        if (unclusteredData.length === 1) {
+          mount('detail', <Detail data={unclusteredData[0]} />);
+          flyToLocation(unclusteredData[0].longitude, unclusteredData[0].latitude);
+        }
+      }
+    };
     mapLocation?.addSource('pinPoint', {
       type: 'geojson',
       data: {
@@ -68,14 +76,5 @@ export const globeCluster = ({ mapLocation, postsData, mount, flyToLocation, pos
 
   if (postModalOpen) {
     removeMapLayersAndSource(mapLocation);
-  }
-
-  function handleUnclusteredPointClick(e: any) {
-    if (postsData) {
-      const postId = e.features[0].properties.cluster;
-      const unclusteredData = postsData.filter(item => item.id === postId);
-      mount('detail', <Detail data={unclusteredData[0]} />);
-      flyToLocation(unclusteredData[0].longitude, unclusteredData[0].latitude);
-    }
   }
 };
